fix: persist sessions for proxy-authenticated users
- Issue Aurral sessions from trusted proxy identities
- Remove reauth recovery navigation and preserve proxy logout behavior
- Update frontend tests and reverse-proxy documentation
