+++ docs/src/content/docs/api/endpoints.mdx
@@ -26,7 +26,6 @@ can change Aurral, Lidarr, files, users, or external services.
 | `GET` | `/api/health/ws` | WebSocket connection statistics |
 | `POST` | `/api/health/stream-token` | Issue a short-lived token for media streams |
 | `POST` | `/api/auth/login` | Create a password-login session |
-| `GET` | `/api/auth/reauth` | Restart reverse-proxy authentication |
 | `POST` | `/api/auth/logout` | End a login session |
 | `GET` | `/api/auth/me` | Current authenticated identity |
 | `GET` | `/api/auth/api-key` | Read or create the instance API key |
