+++ frontend/src/contexts/AuthContext.jsx
@@ -11,12 +11,6 @@ import {
   loginApi,
   logoutApi,
 } from "../utils/api/endpoints/auth.js";
-import {
-  beginLogoutNavigation,
-  endLogoutNavigation,
-  isProxyAuthActive,
-  syncProxyAuthFromBootstrap,
-} from "../utils/authRecovery.js";
 
 const AuthContext = createContext(null);
 
@@ -46,14 +40,12 @@ export const AuthProvider = ({ children }) => {
       setAuthRequired(isRequired);
 
       if (isRequired && bootstrap.user) {
-        syncProxyAuthFromBootstrap(bootstrap);
+        if (bootstrap.token) setStoredAuth({ token: bootstrap.token });
         setUser(bootstrap.user);
         setIsAuthenticated(true);
         setIsLoading(false);
         return;
       }
-
-      syncProxyAuthFromBootstrap();
 
       if (!isRequired) {
         setUser(
@@ -123,25 +115,20 @@ export const AuthProvider = ({ children }) => {
   }, []);
 
   const logout = useCallback(async () => {
-    beginLogoutNavigation();
+    const proxyLogoutUrl = bootstrap?.proxyLogoutUrl;
+    if (proxyLogoutUrl) {
+      void logoutApi().catch(() => {});
+      clearAuthStorage();
+      invalidateBootstrapCache();
+      window.location.href = proxyLogoutUrl;
+      return;
+    }
+
     try {
       await logoutApi();
     } catch {}
     clearAuthStorage();
     invalidateBootstrapCache();
-
-    let proxyLogoutUrl = bootstrap?.proxyLogoutUrl;
-    if (!proxyLogoutUrl && isProxyAuthActive()) {
-      try {
-        proxyLogoutUrl = (await getBootstrapStatus())?.proxyLogoutUrl;
-      } catch {}
-    }
-
-    if (proxyLogoutUrl) {
-      window.location.href = proxyLogoutUrl;
-      return;
-    }
-    endLogoutNavigation();
     setIsAuthenticated(false);
     setUser(null);
   }, [bootstrap]);
@@ -152,6 +139,8 @@ export const AuthProvider = ({ children }) => {
     return !!user.permissions?.[perm];
   }, [user]);
 
+  const canLogOut = !bootstrap?.proxyAuthEnabled || !!bootstrap?.proxyLogoutUrl;
+
   return (
     <AuthContext.Provider
       value={useMemo(() => ({
@@ -161,11 +150,12 @@ export const AuthProvider = ({ children }) => {
         bootstrap,
         login,
         logout,
+        canLogOut,
         authRequired,
         onboardingRequired,
         refreshAuth: checkAuthStatus,
         hasPermission,
-      }), [isAuthenticated, isLoading, user, bootstrap, login, logout, authRequired, onboardingRequired, checkAuthStatus, hasPermission])}
+      }), [isAuthenticated, isLoading, user, bootstrap, login, logout, canLogOut, authRequired, onboardingRequired, checkAuthStatus, hasPermission])}
     >
       {children}
     </AuthContext.Provider>
